Move membership tcp to a separate package
